Added folder structure to organize src directory
