+++ client/index.jsx
@@ -1,9 +1,6 @@
 // const React = require('react');
 // const ReactDom = require('react-dom');
 import App from './src/app.jsx'
-import ReviewList from './src/reviewList.jsx'
-
-
 
 ReactDOM.render(
   <App/>,
