+++ client/src/reviews/appReview.jsx
@@ -1,8 +1,8 @@
-import Rotate from './sectionBox.jsx'
+import Rotate from '../sectionBox.jsx'
 import styled, {keyframes} from 'styled-components'
-import Description from './description.jsx'
+import ReviewParent from './reviewParent.jsx'
 
-class AppDescription extends React.Component {
+class AppReview extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -17,22 +17,22 @@ class AppDescription extends React.Component {
       isStarted: true,
       isOpen: !this.state.isOpen
     })
+    console.log(this.state.Open)
   }
 
   render() {
-    const DescriptionParentDiv = styled.div`
+    const ReviewParentDiv = styled.div`
       display: block;
       width: 100%;
       min-height: 70px;
       vertical-align: center;
       border-top: 1px solid grey;
-
       :hover {
         background-color: ${props => this.state.isOpen ? "white" : "rgb(230, 240, 255)"};
       }`;
       const fadeIn = keyframes`
       from {
-        transform: translate(0px,-20px);
+        transform: translate(0px,-50px);
         opacity: 0;
       }
 
@@ -51,14 +51,15 @@ class AppDescription extends React.Component {
 
     return(
       <div>
-        <DescriptionParentDiv>
-        <h5 onClick={this.toggleOpen}><Rotate isOpen={this.state.isOpen} isStarted={this.state.isStarted} >+</Rotate> Product Description</h5>
+        <ReviewParentDiv>
+        <h5 onClick={this.toggleOpen}><Rotate isOpen={this.state.isOpen} isStarted={this.state.isStarted} >+</Rotate> Reviews</h5>
         <FadeInDiv>
-        {this.state.isStarted ? (this.state.isOpen ? <Description /> : "") : "" }
-    </FadeInDiv></DescriptionParentDiv>
+        {this.state.isStarted ? (this.state.isOpen ? <ReviewParent /> : "") : "" }
+    </FadeInDiv></ReviewParentDiv>
+
       </div>
     )
   }
 }
 
-export default AppDescription
+export default AppReview
